requirements.txt
python-dotenv==1.0.1
jupyterlab==4.1.8
transformers==4.40.1
